requirements/base.txt
Django>=1.8.9,<1.9.0
Pillow>=3.0.0
psycopg2>=2.6.1
django-imagekit>=3.3
django-tinymce>=2.2.0
django-filebrowser-no-grappelli>=3.6.1
django-modeltranslation>=0.10.2
django-registration>=2.0.4
django-uuslug>=1.1.7
requests>=2.9.1
pycrypto>=2.6.1
djlime>=0.0.35
solid_i18n>=1.1.1